Moved activities to their own package
small bug fix on gameInterface too, resetBoard wasn't working
